CIB-2365: review of the working copy snapshot functionality - removed build option to retain working copy - removed has work dir column from the BUILD_RESULT table - remove the option to clean up working copy snapshots - added upgrade task to cleanup any existing snapshots since they will no longer be supported - removed all associated references to working copies including the working copy page, links, capturing code - added a link to the project home page to allow viewing of the working copy.  A working copy is now the persistent work area for a project.  Projects that run clean builds do not have a working copy - added a new 'browser' for navigating the working copy that supports refreshing nodes and a single close button - removed some dead code. - updated some of the vfs code as part of testing.  It is now possible to browse the pulse:/// file system in a meaning ful way, particularly useful for testing the various sets of nodes - added support to the AgentService for listing recipe base directories.

+++ com.zutubi.pulse.master/src/java/com/zutubi/pulse/master/hibernate/SchemaRefactor.java
@@ -204,10 +204,10 @@ public class SchemaRefactor
      * NOTE: The mapping definition should represent what you want to column to look like AFTER the refresh.  This is
      * a little different from the other methods which need to know what the mappings are BEFORE the refactor.
      *
-     * @param tableName
-     * @param columnName
+     * @param tableName     the name of the table containing the column to be refreshed
+     * @param columnName    the name of the column to be refreshed
      * 
-     * @throws SQLException
+     * @throws SQLException is thrown on error
      */
     public void refreshColumn(final String tableName, final String columnName) throws SQLException
     {
@@ -249,10 +249,13 @@ public class SchemaRefactor
         {
             public Object execute(Connection con) throws SQLException
             {
-                Table table = getTable(tableName);
-                Column column = getColumn(table, columnName);
-                dropColumnConstraints(con, table, column);
-                sqlDropColumn(con, table, columnName);
+                if (JDBCUtils.columnExists(con,  tableName, columnName))
+                {
+                    Table table = getTable(tableName);
+                    Column column = getColumn(table, columnName);
+                    dropColumnConstraints(con, table, column);
+                    sqlDropColumn(con, table, columnName);
+                }
                 return null;
             }
         });
@@ -264,8 +267,11 @@ public class SchemaRefactor
         {
             public Object execute(Connection con) throws SQLException
             {
-                Table table = getTable(tableName);
-                dropTable(con, table);
+                if (JDBCUtils.tableExists(con, tableName))
+                {
+                    Table table = getTable(tableName);
+                    dropTable(con, table);
+                }
                 return null;
             }
         });
@@ -362,10 +368,10 @@ public class SchemaRefactor
      * Executes a set of alter table statements to synchronize the underlying database table with the hibernate schema.
      * NOTE: This will only add columns that do not already exist.
      * 
-     * @param table
-     * @param connection
+     * @param table         the table to be altered
+     * @param connection    the database connection providing access to the table
      * 
-     * @throws SQLException
+     * @throws SQLException is thrown on error
      */
     private void updateTableSchema(Table table, Connection connection) throws SQLException
     {
